fix: Rename `ModifySession`
