+++ src/ai/backend/manager/services/session/actions/modify_session.py
